feat: add auth in admin page path
