+++ .env.example
@@ -1,2 +1,3 @@
 NEXT_PUBLIC_BASE_URL=
 OVERCLICKED_DB_DIR=.
+OVERCLICKED_PASSWORD=changeme
