+++ src/app/[password]/admin/[register]/page.tsx
