+++ src/app/[password]/admin/[register]/public/page.tsx
